+++ src/views/Welcome.cpp
@@ -13,8 +13,9 @@ void Welcome::contents() {
     iconsTZE = ui->buttons->containedDraw("Icons", 10, 200, selector == 3, ICON_MDI_ICE_CREAM);
     textsTZE = ui->buttons->containedDraw("Texts", 10, 270, selector == 4, ICON_MDI_FORMAT_FONT);
     imeTZE = ui->buttons->containedDraw("IME", 10, 340, selector == 5, ICON_MDI_KEYBOARD);
+    checkboxesTZE = ui->buttons->containedDraw("Checkboxes", 10, 410, selector == 6, ICON_MDI_CHECK);
 
-    exitTZE = ui->buttons->containedDraw("Exit", 800, 480, selector == 6);
+    exitTZE = ui->buttons->containedDraw("Exit", 800, 480, selector == NUMBER_OF_BUTTONS);
 }
 
 void Welcome::controls() {
@@ -47,6 +48,11 @@ void Welcome::controls() {
         viewsController->setActualView("ImeSample");
     }
 
+    if (ui->buttons->onTouch(checkboxesTZE, utils->touch->lastClickPoint) ||
+        ui->buttons->onPad(checkboxesTZE, utils->pad->pressed.cross)) {
+        viewsController->setActualView("CheckboxesSample");
+    }
+
     if (ui->buttons->onTouch(exitTZE, utils->touch->lastClickPoint) ||
         ui->buttons->onPad(exitTZE, utils->pad->pressed.cross)) {
         viewsController->setActualView(VIEWS_CONTROLLER_EXIT);
@@ -59,10 +65,10 @@ void Welcome::controls() {
     if (utils->pad->pressed.up) {
         selector--;
     }
-    if (selector > 6)
+    if (selector > NUMBER_OF_BUTTONS)
         selector = 1;
     if (selector == 0)
-        selector = 6;
+        selector = NUMBER_OF_BUTTONS;
 
     //touch & pad switch
     if (utils->touch->clicking) {